Error handling framework for catching exceptions, currentyl checking duplicate users, user not found, and bad input

+++ smartsync-user-service/src/main/java/com/smartsync/service/UserService.java
@@ -60,9 +60,18 @@ public class UserService {
 
     public User deleteUser(String id) {
         User deletedUser = this.userRepository.findByUserId(id);
-        this.userRepository.delete(id);
 
-        return deletedUser;
+        if(deletedUser == null) {
+            return null;
+        }
+
+        else {
+            this.userRepository.delete(id);
+
+            return deletedUser;
+        }
+
+
     }
 
 }
